requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
opencv-python==4.10.0.84
numpy==2.1.2
pillow==11.0.0
qdrant-client==1.12.1
python-multipart==0.0.12